requirements.txt
aiofiles==23.1.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
install==1.3.5
lxml==4.9.3
outcome==1.2.0
packaging==23.1
pyaes==1.6.1
pyasn1==0.5.0
Pyrogram==2.0.106
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
rsa==4.9
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
Telethon==1.28.5
temp==2020.7.2
TgCrypto==1.2.5
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
